fix(security): complete I9-I18 hardening
