+++ src/features/plans/lifecycle/generation.ts
@@ -112,7 +112,7 @@ async function runGeneration(
       status: 'failure',
       classification: result.classification,
       error: result.error,
-      metadata: result.metadata as Record<string, unknown> | undefined,
+      metadata: result.metadata,
       usage: result.metadata ? safeNormalizeUsage(result.metadata) : undefined,
       durationMs: result.durationMs,
       timedOut: result.timedOut,
@@ -128,7 +128,7 @@ async function runGeneration(
       status: 'failure',
       classification: exec.classification,
       error: exec.error,
-      metadata: exec.metadata as Record<string, unknown> | undefined,
+      metadata: exec.metadata,
       usage: exec.metadata ? safeNormalizeUsage(exec.metadata) : undefined,
       durationMs: exec.durationMs,
       reservation: exec.reservation,
@@ -140,7 +140,7 @@ async function runGeneration(
   return {
     status: 'success',
     modules: exec.modules as GeneratedModule[],
-    metadata: exec.metadata as Record<string, unknown>,
+    metadata: exec.metadata,
     usage: safeNormalizeUsage(exec.metadata),
     durationMs: exec.durationMs,
     reservation: exec.reservation,
