+++ supabase/schema/tables/tasks.ts
@@ -311,13 +311,6 @@ export const taskProgress = pgTable(
       `,
       }),
 
-      // Defense in depth only: authenticated DELETE is revoked at the table level;
-      // this ownership policy remains inert unless DELETE is deliberately reintroduced.
-      pgPolicy('task_progress_delete_own', {
-        for: 'delete',
-        to: 'authenticated',
-        using: userOwnsRecord,
-      }),
     ];
   },
 ).enableRLS();
